fix ambiguous sin in unit test for clang compatibility
Object converts implicitly to bool/int/double, so unqualified sin(d[0])
is ambiguous under clang (MSVC picks the double overload silently).
Convert to double explicitly, matching the native-branch std::sin call.

+++ unit_test/cifa_unit_test.cpp
@@ -111,7 +111,7 @@ bool register_function_test()
         c1.register_native_function("sin", [](TestBytecode::NativeCallContext& context)
             { context.set_result(std::sin(context.to_number(0))); });
     else
-        c1.register_function("sin", [](ObjectVector& d) { return sin(d[0]); });
+        c1.register_function("sin", [](ObjectVector& d) { return std::sin(d[0].toDouble()); });
 
     std::string script_code = R"(
     double PI = 3.141592653589793238462643383279;
